agent_stack/requirements.txt
numpy
requests
sentence-transformers
faiss-cpu
tqdm
pypdf2
google-genai
flask